requirements.txt
# Core
numpy
pandas
tqdm
pyyaml

# Text & NLP
sentence-transformers
langchain
faiss-cpu
chromadb
openai
tiktoken

# PDF & file handling
PyPDF2
unstructured
pdfplumber
feedparser

# App
streamlit
matplotlib
plotly

# Utilities
loguru
